Move default `players` list from `src/App.vue` to `src/data/players.ts` and import it to `App`
